put #ifndef _OPTIMIZED_FUNCTIONS around the functions using reorder_* in evaluateGenericSpecial.c and makenewzGenericSpecial.c.
This means that now the code can be compiled without -O2, as long as _OPTIMIZED_FUNCTIONS is defined. If it is not defined (i.e., the 'unotimized' functions that currently contain the reorder_* calls are used, there will be linker errors regardless of the optimization flags, because the reorder functions are missing (they are implemented in sim_reorder.cpp, which is currently _not_ compiled).