added macro RAXML_USE_LLALLOC to switch on the lockless_allocator (the code in mem_alloc.c will dispatch to the actual malloc implementations). If RAXML_USE_LLALLOC is defined, it is also necessary to link lockless_allocator/ll_alloc.c.
When RAXML_USE_LLALLOC is defined, it is a good idea to run the check_malloc.sh script on the resulting binary to prevent any nastiness (e.g., killing kittens) caused by accidental use of the default allocator functions.

I've tried to modify the SSE/AVX/PTHREADS makefiles accordingly, but have not touched the generic makefiles, yet.
Lockless alloc is only activated in the PTHREADS makefiles, because it is not necessary for single threaded runs.

In any case it is now necessary to link mem_alloc.c where the forwarding of the rax_* functions is done. It is not possible to do this 'header-only' with static functions, because this would cause tons of defined-but-not-used warnings.